Code cleanup and split dashboard scss file into two component-related files
