+++ next/src/app/dashboard/weekly_overview.module.scss
@@ -1,0 +1,120 @@
+@use 'src/app/globalutil';
+
+.weekdays {
+  width: 100%;
+  min-width: 800px;
+  
+  height: 35px;
+  display: flex;
+  justify-content: space-around;
+  text-align: center;
+
+  flex: 0 0 35px;
+
+  p:last-child {
+    font-weight: 300;
+    font-size: 0.7em;
+    margin-top: 2px;
+    color: rgb(160, 157, 157);
+  }
+}
+
+
+/* Hourly breakdown */
+.hourlywrapper {
+  height: 100%;
+  width: 100%;
+  margin-top: 7px;
+
+  min-width: 800px;
+
+  overflow: hidden;
+}
+.hourly {
+  display: flex;
+
+  height: 100%;
+  width: 100%;
+  max-height: 100%;
+
+  overflow-y: auto;
+  scrollbar-width: none;
+  -ms-overflow-style: none;
+  scrollbar-width: thin;
+}
+
+/* Hours Section */
+.hours {
+  width: 100%;
+  height: fit-content;
+
+  border: 1px solid var(--dividers);
+  border-right: none;
+
+  overflow: hidden;
+
+  &:last-child {
+    border-right: 1px solid var(--dividers);
+  }
+}
+
+.block {
+  height: 15px;
+  border-bottom: 1px solid #E7E7E7;
+  overflow-y: visible;
+
+  position: relative;
+
+  p {
+    font-size: 10px;
+    font-weight: 400;
+    color: var(--font-color-alt);
+    padding: 2px;
+  }
+
+  &:last-child {
+    border-bottom: none;
+  }
+}
+
+.cover {
+  position: absolute;
+  width: calc(100% + 1px);
+  height: calc(100% + 1px);
+
+  background-color: #e1e1e1;
+  opacity: 0.9;
+  background-size: 5px 5px;
+  background-image: repeating-linear-gradient(-45deg, #aeaeae 0, #aeaeae 1px, #e1e1e1 0, #e1e1e1 50%);
+
+  z-index: 5;
+
+  top: -1px;
+  left: -1px;
+}
+
+.weekly_app {
+  position: absolute;
+  left: 30px;
+  top: 1px;
+
+  padding: 2px;
+
+  line-height: 0.9;
+
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+
+  width: calc(100% - 30px);
+  background-color: #BA1682;
+
+  z-index: 1;
+
+  p {
+    color: var(--font-on-dark);
+    font-size: 10px;
+  }
+
+  @include globalutil.deep_shadow;
+}
